util/scripts/requirements.txt
numpy-1.13.1
pandas-0.20.3
pysam-0.14.1
